Convert capule to submodule

+++ data/include/telecom.h
@@ -16,12 +16,12 @@ public:
     void reset_cmd();
 
     static void handle_uplink(int packet_size);
-    void handle_capsule_uplink(uint8_t packet_id, uint8_t* data_in, uint16_t len);
+    void handle_capsule_uplink(uint8_t packet_id, uint8_t* data_in, uint32_t len);
     static void handle_downlink(int packet_size);
-    void handle_capsule_downlink(uint8_t packet_id, uint8_t* data_in, uint16_t len);
+    void handle_capsule_downlink(uint8_t packet_id, uint8_t* data_in, uint32_t len);
 
 private:
-    void send_packet(uint8_t packet_id, uint8_t* data, uint16_t len);
+    void send_packet(uint8_t packet_id, uint8_t* data, uint32_t len);
 
     Capsule<Telecom> capsule_uplink;
     Capsule<Telecom> capsule_downlink;
